requirements.txt
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
fastapi>=0.110.0
uvicorn>=0.29.0
pyyaml>=6.0
openpyxl>=3.1.0
xlrd>=2.0.1
pyxlsb>=1.0.10
python-multipart>=0.0.9
